initial async support
This is based on the unstable hyper master branch (which is intended to become
hyper 0.11.x). I'm likely to tweak it still; in particular, it might be better
to use futures to defer the calls to Entity::get_range, which would require
accepting Entity as something like a Deref<Entity> + Send + 'static rather
than the current &Entity.