static file serving, with example
I changed the http_entity::serve interface a bit in the process:

   * return a Response rather than a FutureResult<Response, Error>.
     It was always returning future::ok anyway. This simplifies
     returning it from a CpuFuture; having one future which returns
     another of a different kind would be needlessly complicated.

   * take an Entity rather than a Deref<Entity> and require Entity to
     be Send. File needs to be defined as struct File(Arc<FileInner>)
     so that it can both pass the FileInner to the CpuPool and use it
     for the spawn call. Having a Box/Arc of the Arc is wasteful.
     Passing the Entity by value avoids this. This should be a strict
     superset of the old approach; anything that was possible before
     should still be possible by using a wrapper type.

And I fixed a bug: the etag and last-modified weren't being returned in
multi-range responses.

File serving might be doing more thread handoffs than necessary; in
particular, even though the serve call is happening on the CpuPool, it
enqueues the first chunk on the pool rather than starting it eagerly,
and I wonder if that means it doesn't start until there's a context
switch to the reactor and back. Maybe that will be the focus of a future
commit, though I'm not sure now how I'd fix it cleanly.