speed up file serving
This was terribly slow before when serving small files, and I'd guessed
that it was doing extra bouncing between the reactor and the cpu pool on
every chunk. The new method uses the mpsc to transfer between rather
than creating a new CpuFuture from the reactor every time. This appears
significantly faster.

Note that when constrained to a single cpu via "taskset --cpu-list 0",
the difference is not nearly as pronounced. This probably supports the
idea the problem is synchronization-related.

Approximate wrk results (default parameters, just one sample):
Before:  860 requests per second
After:  6500 requests per second