single crate (http-serve)
The dependencies of http-file weren't that different from those of
http-entity. The only addition was futures-cpupool, and hyper depends on that
anyway, so there's no net addition of transitive dependencies. So just be
simple and combine them.

The name http-entity was bad anyway---it's the HTTP RFCs' name for the core
abstraction but doesn't actually say what you can do with it. Hopefully
http-serve is a more clear.